Tests/utils: renaming utils as common as menionned in rust book
